+++ bugzoo/mgr/bug.py
@@ -56,11 +56,17 @@ class BugManager(object):
 
     add = register
 
+    def __delitem__(self, name: str) -> None:
+        """
+        Attempts to deregister a bug with a given name.
+        """
+        del self.__bugs[name]
+
     def deregister(self, bug: Bug) -> None:
         """
         Attempts to deregister a given bug from this manager.
         """
-        del self.__bugs[bug.name]
+        del self[bug.name]
 
     remove = deregister